feature-content-curator - Added Content Curator Role that allows user to see filter or non filtered output
