+++ ContentConsole.Model/Constants.cs
@@ -8,5 +8,7 @@ namespace ContentConsole.Model
     {
         public const string INPUT = "The weather in Manchester in winter is bad. It rains all the time - it must be horrible for people visiting.";
         public const string ADMIN_ROLE = "A";
+        public const string CONTENT_CURATOR_ROLE = "C";
+        public const string USER_ROLE = "U";
     }
 }
